add Map && upload image

+++ app/src/main/java/ir/co/tarhim/model/deceased/CreateDeceasedRequest.kt
@@ -1,5 +1,7 @@
 package ir.co.tarhim.model.deceased
 
+import java.nio.channels.FileLock
+
 data class CreateDeceasedRequest(
     val accesstype: String,
     val birthday: String,
@@ -7,8 +9,8 @@ data class CreateDeceasedRequest(
     val deathloc: String,
     val description: String,
     val imageurl: String,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val name: String
 
 )
